+++ app/models/settings_template.rb
@@ -3,7 +3,7 @@ class SettingsTemplate < ActiveRecord::Base
 
   attr_accessible :name, :intouch_settings
 
-  store :intouch_settings, accessors: %w(assigner_groups reminder_settings telegram_settings email_settings)
+  store :intouch_settings, accessors: %w[assigner_groups assigner_roles reminder_settings telegram_settings email_settings]
 
   def copy_from(settings_template)
     self.attributes = settings_template.attributes.dup.except('id', 'created_on', 'updated_on')